requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.0
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
google-auth==2.40.3
google-genai==1.36.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
torch==2.8.0
torchvision==0.23.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
zipp==3.23.0
